constexpr (almost) all the things

+++ cpplox/src/memory.cpp
@@ -107,7 +107,7 @@ static void markRoots()
 
 static void markArray(ValueArray* array)
 {
-  for (int i = 0; i < array->count; i++) {
+  for (size_t i = 0; i < array->count; i++) {
     markValue(array->values[i]);
   }
 }
